requirements.txt
numpy>=1.26,<2.0
opencv-python>=4.8.1.78
pandas>=2.2.3
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
Pillow>=10.2.0,<10.3.0
scikit-learn>=1.5.2
pytest>=7.4.3
facenet-pytorch==2.6.0
openpyxl>=3.1.4


